python_analysis/requirements.txt
# Core Reddit and NLP
praw==7.7.1
python-dotenv==1.0.0
textblob==0.17.1

# Data processing (compatible versions)
pandas>=2.0.0
numpy>=1.24.0

# ML and clustering (use newer compatible versions)
sentence-transformers>=2.2.0
scikit-learn>=1.4.0
umap-learn>=0.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0
plotly>=5.15.0

# Optional AI (if needed)
openai>=1.0.0
